Fix useless [event] in [titlescreen_music]
The default playlist sets the music list twice, once before the
story screen and again in prestart, but the order of tracks is different.

This means that the track order at the titlescreen is now the same
as the track order in a scenario using the default playlist,
which previously was not the case.

Whether that actually matters? I have no idea. I doubt it.

In any case, as of this commit, there are no schema validation errors
at the title screen.