EI: Fix issues found by the schema validator
With this, EI is as clean of schema errors as it's going to get.
There yet remain a number of unsolvable schema errors (due to multi-use macros).

Regarding {LOYAL_UNIT}, specifying the side here is redundant;
however, an empty side key raises an error, and it doesn't hurt
to specify the redundant information.